extract texera_filter_operator.py
